Add headers abstract class
